feat/13: move convertScheduleToTaskData to separate file. account for null date.
